+++ src/cprunner/__main__.py
@@ -1,0 +1,5 @@
+import run
+
+
+if __name__ == "__main__":
+    run.executor()
